+++ packages/vaex-ml/vaex/ml/transformations.py
@@ -32,6 +32,7 @@ class Transformer(HasState):
     ''' Parent class for all of the transformers.
     '''
     features = traitlets.List(traitlets.Unicode(), help=help_features).tag(ui='SelectMultiple')
+    features_ = traitlets.List(traitlets.Unicode()).tag(output=True)
 
     def fit_transform(self, df):
         '''Fit and apply the transformer to the supplied DataFrame.
@@ -42,6 +43,9 @@ class Transformer(HasState):
         '''
         self.fit(df=df)
         return self.transform(df=df)
+
+    def _get_output_features(self):
+        self.features_ = [self.prefix + feat for feat in self.features]
 
 
 @register
@@ -100,6 +104,8 @@ class PCA(Transformer):
         self.explained_variance_ = self.eigen_values_
         self.explained_variance_ratio_ = (eigen_values[indices] / np.sum(eigen_values)).tolist()
 
+        self._get_output_features(df)
+
     def transform(self, df, n_components=None):
         '''Apply the PCA transformation to the DataFrame.
 
@@ -127,6 +133,13 @@ class PCA(Transformer):
             name = self.prefix + str(i + name_prefix_offset)
             copy[name] = expr
         return copy
+
+    def _get_output_features(self, df):
+        name_prefix_offset = 0
+        while self.prefix + str(name_prefix_offset) in df.get_column_names(virtual=True, strings=True):
+            name_prefix_offset += 1
+
+        self.features_ = [self.prefix + str(i + name_prefix_offset) for i in range(self.n_components)]
 
 
 @register
@@ -237,6 +250,8 @@ class LabelEncoder(Transformer):
             labels = vaex.array_types.tolist(df[feature].unique())
             self.labels_[feature] = dict(zip(labels, np.arange(len(labels))))
 
+        self._get_output_features()
+
     def transform(self, df):
         '''
         Transform a DataFrame with a fitted LabelEncoder.
@@ -285,9 +300,8 @@ class OneHotEncoder(Transformer):
      3  blue                1              0            0
      4  red                 0              0            1
     '''
-
     # title = Unicode(default_value='One-Hot Encoder', read_only=True).tag(ui='HTML')
-    prefix = traitlets.Unicode(default_value='', help=help_prefix).tag(ui='Text')
+    prefix = traitlets.Unicode(default_value='', help=help_prefix)
     one = traitlets.Any(1, help='Value to encode when a category is present.')
     zero = traitlets.Any(0, help='Value to encode when category is absent.')
     uniques_ = traitlets.List(traitlets.List(), help='The unique elements found in each feature.').tag(output=True)
@@ -312,6 +326,8 @@ class OneHotEncoder(Transformer):
             uniques.append(unique_values)
         self.uniques_ = uniques
 
+        self._get_output_features()
+
     def transform(self, df):
         '''Transform a DataFrame with a fitted OneHotEncoder.
 
@@ -332,6 +348,9 @@ class OneHotEncoder(Transformer):
                 else:
                     copy[column_name] = copy.func.where(copy[feature] == value, self.one, self.zero)
         return copy
+
+    def _get_output_features(self):
+        self.features_ = [self.prefix + feat + '_' + str(value) for i, feat in enumerate(self.features) for value in self.uniques_[i]]
 
 
 @register
@@ -375,6 +394,8 @@ class FrequencyEncoder(Transformer):
         # Encoding
         for feature in self.features:
             self.mappings_[feature] = dict(df[feature].value_counts() / nsamples)
+
+        self._get_output_features()
 
     def transform(self, df):
         '''Transform a DataFrame with a fitted FrequencyEncoder.
@@ -441,6 +462,8 @@ class StandardScaler(Transformer):
 
         assign(mean, std)
         df.execute()
+
+        self._get_output_features()
 
     def transform(self, df):
         '''
@@ -515,6 +538,8 @@ class MinMaxScaler(Transformer):
         assign(minmax)
         df.execute()
 
+        self._get_output_features()
+
     def transform(self, df):
         '''
         Transform a DataFrame with a fitted MinMaxScaler.
@@ -575,6 +600,8 @@ class MaxAbsScaler(Transformer):
         absmax = df.max(['abs(%s)' % k for k in self.features]).tolist()
         # Check if the absmax_ value is 0, in which case replace with 1
         self.absmax_ = [value if value != 0 else 1 for value in absmax]
+
+        self._get_output_features()
 
     def transform(self, df):
         '''
@@ -648,6 +675,8 @@ class RobustScaler(Transformer):
         if self.with_scaling:
             self.scale_ = (df.percentile_approx(expression=self.features, percentage=q_max) - df.percentile_approx(expression=self.features, percentage=q_min)).tolist()
 
+        self._get_output_features()
+
     def transform(self, df):
         '''
         Transform a DataFrame with a fitted RobustScaler.
@@ -708,7 +737,8 @@ class CycleTransformer(Transformer):
 
         :param df: A vaex DataFrame.
         '''
-        pass
+
+        self._get_output_features()
 
     def transform(self, df):
         '''
@@ -724,6 +754,11 @@ class CycleTransformer(Transformer):
             copy[name_y] = np.sin(2 * np.pi * copy[feature] / self.n)
 
         return copy
+
+    def _get_output_features(self):
+        for feature in self.features:
+            self.features_.append(self.prefix_x + feature + self.suffix_x)
+            self.features_.append(self.prefix_y + feature + self.suffix_y)
 
 
 @register
@@ -778,6 +813,8 @@ class BayesianTargetEncoder(Transformer):
             agg = df.groupby(feature, agg={'count': vaex.agg.count(), 'mean': vaex.agg.mean(self.target)})
             agg['encoding'] = (agg['count'] * agg['mean'] + self.weight * global_target_mean) / (agg['count'] + self.weight)
             self.mappings_[feature] = {value[feature]: value['encoding'] for index, value in agg.iterrows()}
+
+        self._get_output_features()
 
     def transform(self, df):
         '''Transform a DataFrame with a fitted BayesianTargetEncoder.
@@ -856,6 +893,8 @@ class WeightOfEvidenceEncoder(Transformer):
             agg['negative'] = agg.func.where(agg['negative'] == 0, self.epsilon, agg['negative'])
             agg['woe'] = np.log(agg.positive/agg.negative)
             self.mappings_[feature] = {value[feature]: value['woe'] for index, value in agg.iterrows()}
+
+        self._get_output_features()
 
     def transform(self, df):
         '''Transform a DataFrame with a fitted WeightOfEvidenceEncoder.