+++ CHANGELOG.md
@@ -212,6 +212,7 @@
       * BayesianTargetEncoder [#533](https://github.com/vaexio/vaex/pull/533)
       * WeightOfEvidenceEncoder [#534](https://github.com/vaexio/vaex/pull/534)
       * Improve the consistency of the vaex.ml API for model transformers [#552](https://github.com/vaexio/vaex/pull/552)
+      * Add `features_` property (list of output features) to the base Transformer class [#553](https://github.com/vaexio/vaex/pull/553)
 
 # vaex 2.5.0 (2019-12-16)